reinterpert_cast replaced by a C-style cast, to avoid compiler errors on some plataforms.
Cast an object pointer to a function pointer is an undefinied behaviour in some exotic
platforms, so some compiler raise a flag against it. However if we use a C-style cast the compiler
ignores it.

This problem is related to the QLibrary API returning a void* instead of a generic
function pointer.