APIExtractor is a huge amount of legacy code inherited from QtScriptGenerator, but QtScriptGenerator itself isn't a library, this explains why libapiextractor does not care about things that every library SHOULD care, symbol visibility and binary compatibility.
This commit adds symbol visibility rules to libapiextractor as the first step to make
libapiextractor aware of binary compatibility.

This is also needed if we want to be able to compile and use libapiextractor
under Windows.

Note: Not all symbols were made public, just the symbols needed by shiboken,
boostpython and doc generators, because IMHO libapiextractor needs some love
and a API review. More symbols could be added later if needed.

Reviewed by Renato Araujo <renato.filho@openbossa.org>