Added the method injectedCodeCallsPythonOverride(func) to ShibokenGenerator, it returns true if a code injection on a virtual method override (a "native" inject-code) does a call to the Python override.
Added the %PYTHON_METHOD_OVERRIDE type system variable,
it is replaced by the name of the pointer to a Python
method that represents an override to a C++ virtual method.

A code injection in the "native/end" position for a method
modification is now put before the dereferencing of the
variables used in the Python call (the method object and
the Python argument tuple).

If a call to the Python override method is detected on
code injections of the "native" class from method
modifications, the generator doesn't write the same
call again.

All documentation was updated with the changes.