Added the "containsCppWrapper" flag to the PyBaseWrapper.
The containsCppWrapper flag says if the C++ pointer held by the
Python wrapper has a binding generated C++ wrapper and was created
by Python. This means that the C++ object has a virtual destructor
that will invalidate the Python wrapper if it is destroyed when its
ownership is with C++.

PyBaseWrapper_New function was expanded with the "containsCppWrapper"
parameter, and CppGenerator::writeConstructorWrapper was updated to
use it.

The PyBaseWrapper structure was documented with comments.