Added method to take care of the details of ownership transfer to C++.
Simple ownership transference, i.e. without parenting, is now performed
by the new BindingManager::transferOwnershipToCpp method. It remove the
parent of the transfered object and proceeds transfer or invalidation
if needed. The generated code for simple ownership transfer from Python
to C++ now reflects this change.

Fixed the method BlackBox::keepObjectType that steals an ObjectType
ownership to C++ to remove it from its parent also. The BlackBox
class does not take care of the stolen object as a proper parent
would, but its destructor deletes the object, so the "unparenting"
is needed to avoid freeing the same memory twice.

Created an unit test that adds children to a parent ObjectType and then
steal the children to C++ with BlackBox.keepObjectType.

Reviewed by Hugo Lima <hugo.lima@openbossa.org>