Added Shiboken_TypeCheck macro that makes use of PyType<T>().
The Shiboken_TypeCheck calls Python's PyObject_TypeCheck using the
type pointer stored in PyType<T> for the type being checked.
Conversion<T*>::toCpp(pyobj) converter tries first to convert pyobj
to the Python wrapper for type T and only second to any of the
convertible types. If pyobj is neither of those, 0 is returned as
the C++ object. This works fine for Py_None and invalid values are
not expected to be passed because the generated code checked the
types first.