Converter<T>::copyCppObject method was replaced by a template function.
Added the template function

T* SbkCopyCppObject(const T& cppobj);

whose task is simply to copy a C++ object. If the binding has a C++ class
wrapper for it, the function is specialized to use the copy constructor
of the C++ wrapper class. This replaces the  Converter<T>::copyCppObject
method.

Also moved implementation of Converter<T>::toPython from the generator to
ConverterBase<T> in the conversions header, for it makes use of the
SbkCopyCppObject.

Reviewed by Lauro Neto <lauro.neto@openbossa.org>