Added the option --enable-pyside-extensions to enable PySide extensions.
You need to enable these extensions if you are generating a binding for a Qt-based library
and need to pass some types through the signal slots mechanism.

If you turn this flag on, you also need to link the generated bindings with the
libpyside, besides add libpyside include path to your project.

Reviewed by Marcelo Lira <marcelo.lira@openbossa.org>