Adds normalizedSignature static method to TypeDatabase.
All the uses of QMetaObject::normalizedSignature were replaced to use
TypeDatabase::normalizedSignature; the former always changes the
unsigned primitives to the shorter versions (e.g. "unsigned int"
becomes "uint") that aren't necessarily used by the parsed library.
The new normalizer changes back "uNUMBER" to "unsigned NUMBER" if the
former is not present in the TypeDatabase.

Reviewed by Hugo Parente <hugo.lima@openbossa.org>