Adds test for modified function using type system template.
The new test adds a "toInt(bool* ok = 0, int base = 10)" method to
Str class. The modification uses a type system template function
that uses the variable "%2" to reference the "base" argument; in
the case when the thing is called without parameters "%2" should be
replaced by "10" instead of a converted C++ variable name.

Reviewed by Hugo Parente <hugo.lima@openbossa.org>