Adds support for Duck Punching (aka Monkey Patching).
The simplest definition of duck punching for our purposes is
the ability to change the definition of a method in an instance of
a class.
To allow this behaviour the SbkBaseWrapper structure had to be
extended with a 'ob_dict' which is a PyObject pointer to the instance
dictionary. It is originally set to NULL until the user tries to access
it. This dictionary could be accessed through the '__dict__' instance
property. For now it is read-only.

The generator was updated to handle the instance dictionary, and an
extensive duck punching test was also added.

Reviewed by Hugo Parente Lima <hugo.lima@openbossa.org>