Changed API for PySequence_to_argc_argv.
It was renamed to PySequenceToArgcArgv and a new argument was added, besides some documentation.

If the sequence is empty and defaultAppName (the new argument) was provided, argc will be 1 and
argv will have a copy of defaultAppName because some libraries, like Qt, need at least one element
in argv (the application name), otherwise it'll crash somewhere inside Qt.

Reviewed by Luciano Wolf <luciano.wolf@openbossa.org>