Adds support for managing objects referred by a Python wrapper.
A mapping to referred objects has been added to the SbkBaseWrapper
structure to keep track of objects used by a given Python
wrapper. This differs from ownership or parenting, since
the referee is not responsible for destroying the
referred object. It is more akin to a model/view relationship
when many views refers to one model but don't own it, and
the model must be kept alive as long it is referred.

Two methods were added to assist the reference keeping:

* SbkBaseWrapper_keepReference
  Causes a referred object reference counter to be
  increased and any previous used object refcount is
  decreased.
SbkBaseWrapper_clearReferences
  Decrements the reference counter of all referred
  objects. It is called when the Python wrapper referee
  is destroyed.