The C++-like enum values are now registered inside the related enums.
Originally the values of an enum were registered in the scope that
enclosed the enum declaration, just like C++ does, now in addition
to this the values are registered inside the enum type.
To exemplify, the following C++ enum:

  Scope {
    enum Foo {
        Value
    };
  };

can be accessed in Python as this:

  Scope.Value

as well as this:

  Scope.Foo.Value

The enum unit tests were expanded to check for this new behaviour.

Reviewed by Luciano Wolf <luciano.wolf@openbossa.org>