Remove inplace operators of flags
Enums and flags are now treated as immutable types, so
using a inplace operator will create a new flag object
and attribute it to the name which the operator is called.

flags = Qt.Window
flags |= Qt.Dialog # This will create a new object and attribute
                   # to flags

This was done in order to prevent changing the original objects as
the the inplace operator in line2 would modify the object pointed by
Qt.Window.

Reviewed by Marcelo Lira <marcelo.lira@openbossa.org>