Added method AbstractMetaBuilder::fixReturnTypeOfConversionOperator(func).
The new method fixes the return type of conversion operators:
they should return the target of the conversion as type and not
the type of its owner class.
fixReturnTypeOfConversionOperator is used in traverseFunctions.
An unit test was added for this case.

Note that this behaviour could be fixed in the parser. I dare you!