Adds a test for a method with modifications and custom conversion rules.
The test function does nothing in fact, it is there just to
cause a situation that could cause a compilation problem if
the generator is not working as it should. In other words
if it compiles, then it is ok. :)

The added method was a dummy clone of a similar QPixmap.loadFromData
method signature that prevented QtGui from compiling properly.

Reviewed by Hugo Parente <hugo.lima@openbossa.org>
Reviewed by Anderson Lizardo <anderson.lizardo@openbossa.org>