Fixes attributes for default constructors and class' "isPolymorphic" flag.
When a class has no constructor explicitly declared a default is
provided by API Extractor, but it should be marked as "Final".

The setting of the "isPolymorphic" flag done in the
AbstractMetaClass::addFunction method must take into account if
the class has a virtual destructor, as it is done in the
AbstractMetaClass::setFunctions method.

The already existing tests cover the cases affected by the changes.