Remove export macros from inlined classes.
The GilState and ThreadStateSaver classes are purely inline,
and should not be declared with dllexport/dllimport macros. The
dllimport macro forces MSVC not to inline the classes in client code,
which causes linker errors, since the symbols are not included in the
shiboken dll.

Reviewer: Hugo Parente <hugo.lima@openbossa.org>
Reviewer: Luciano Wolf <luciano.wolf@openbossa.org>