Added methods to check if overloads contain any argument with default values.
The hasArgumentWithDefaultValue methods from OverloadData returns
true if any argument in the overloads or in the single function
passed has a default value assigned to it.
A non-static version of this method uses the overloads described
by the OverloadData object.

Also added the method getArgumentsWithDefaultValues(func) to return
a list of the function's arguments that have default values.