BindingManager destructor now asserts if the wrapper map is empty.
The assertion is only used in debug mode. The duck punching test
was altered to avoid failing when run with debug. Check the
duck punching test file for a better explanation.

Also added a debug helper method "showWrapperMap" to the binding
manager, it shows the contents of the C pointer to Python wrapper
mapping.

Reviewed by Luciano Wolf <luciano.wolf@openbossa.org>
Reviewed by Renato Araújo <renato.filho@openbossa.org>