Revert "Created operator != or __ne__ in Point class, to avoid  tests errors with Python 2.7."
This reverts commit ce4b5f535f43bd2ff800b709f88c6f9c5be68003.

The unittest.assertNotEqual method has changed in Python 2.7 to use the
"!=" operator. The Point class was intentionally left without a "!="
to test if Shiboken would raise a NotImplementedError exception.
This commit will be reverted and the tests using Point and
assertNotEqual must be changed to avoid the possible usage of "!=".