Updated Shiboken cmake files to provide build type information.
The build type information, contained in the the SHIBOKEN_BUILD_TYPE
variable, will tell modules linking with libshiboken if
it was built in Release or Debug mode.

The SHIBOKEN_PYTHON_INTERPRETER variable was also added to the cmake
Shiboken files, to make it known which Python interpreter (normal or
debug, in systems where this separation exists).

SHIBOKEN_PYTHON_LIBRARIES variable tells which Python library
libshiboken was linked against.