Fixed overload decisor sorting to put QStrings after pointers to wrapped objects.
This is a special extension for Qt bindings. Since QStrings accept None
values the same way object and value types accept, to avoid confusion
and calling the wrong signature QString must go after object and value
pointers. This wasn't a problem before, but now QString is a
primitive-type and the decisor has no access to its implicit
conversions, and thus can't sort it properly.

Reviewed by Luciano Wolf <luciano.wolf@openbossa.org>
Reviewed by Renato Araújo <renato.filho@openbossa.org>