Type system parser has now the ability to ignore entries as demanded by the user.
The entries that could be dropped are:
* Object and Value types
* Global functions
* Namespaces
* Enums

The entry name must be fully qualified with scope items separated by a
dot (.) and beginning with the module/package name.
Example: to drop the class "Bar" inside the namespace "Foo" from the
"Pkg" package specify it with: "Pkg.Foo.Bar".

TODO: The parser will later complain that dropped entries found on the
headers are not found in the type system. That's obviously incorrect,
but to fix it all the type entries should store the name of the package
from where the came. And that's a needed improvement!