Added a new test binding called 'minimal'.
The purpose is to have the smallest possible binding to help when doing
complex changes that require a good deal of debugging and tracking,
and the many operations performed in too many wrapped classes will
clutter any output used for those purposes.

In other words: don't add anything here except with a good reason for
it, prefer to use 'sample' or 'other' binding for that.

Reviewed by Hugo Parente <hugo.lima@openbossa.org>
Reviewed by Luciano Wolf <luciano.wolf@openbossa.org>