Included tests for added function signatures very similar to already existing ones.
Specifically this causes the situation when there is in C++ a function
with an argument that is a reference to a Value Type, and the user adds
a very similar function with the same type, but passed as value.

Example:
C++  : function(const TYPE& a, int b)
Added: function(TYPE)

The return type of ShibokenGenerator::getArgumentReplacement() method
was modified, because the argument object is more useful than its index.