Fix and test case for bug PYSIDE-72.
The address of the PyObject is used directly as the hash rather than a
member of that object.  This avoids segfaults when the C++ object has
been deleted, rendering the child pointers NULL.

Removed a test case verifying that hash(QObject()) != hash(QObject())
because they in fact can be equal.  The first QObject dies and is reaped
before the second QObject is created, meaning that the second QObject
may be allocated at the same address as the first QObject, giving them
both the same address.  If a reference is held to the first object,
though, then they will get different hash values.

Change-Id: I116463f88b837726a98720bae36770d53b13f4ee
Reviewed-by: Marcelo Lira <marcelo.lira@openbossa.org>