Improve dependencies for tests
Add the global header and typesystem XML, and shiboken itself, as
dependencies to running Shiboken for generating the wrapped code used in
the unit tests. This gives a better example in case of users looking at
this code for how to do things in their own projects, and also makes
test writing easier by removing the need to do a clean build or touch
files to force the wrapping code to be recreated when modifying the test
code or the generator. Change command invocation to use shiboken as a
target name (CMake will recognize it as a target and use the correct
path automatically), and remove the now redundant dependency of the
wrapping libraries on the same. (This is superfluous both because using
shiboken as a target will cause CMake to add an implicit dependency on
it for users of the generated files, but also because it isn't really
needed; the generate command having it as a dependency covers needing it
built first, and is what we really care about as far as rebuilding
anything if it changes.)

Change-Id: If5cdf2e98602e500b49ba7b000e4e30e77e1a136
Reviewed-by: John Ehresman <jpe@wingware.com>