Remove unnecessary line feed
Remove explicit '\n' from echo command in 'dist' target. This isn't
needed (echo should output one regardless, so it only creates an
additional empty line), and it is being parsed by CMake, which results
in generating a build.ninja on at least Windows that cannot be parsed
due to an unterminated string.

Change-Id: I5d8aba12c71f5a9c29a07108c50499081ba7203a
Reviewed-by: John Ehresman <jpe@wingware.com>