Fix compilation issue on OS X 10.9
Starting with OS X 10.9 the system toolchain defaults to libc++.
This is a problem with the current sparse hash config as it defaults
to using  c++11 functionality under the std::tr1 namespace instead
of std, which fails for libc++ since it does not implement the tr1
namespace. This change allows libc++ based systems to use the correct
namespace.

Change-Id: Ia8dd49fb5fad2915f41e44240e45a7af9d6453a8
Reviewed-by: John Ehresman <jpe@wingware.com>