doc: Stop requiring sphinx.ext.refcounting.
Sphinx 1.2 made a backwards-incompatible change and removed its
`refcounting' extension module because it "is very specific to CPython
and has no place in the main distribution".

Fix the build with Sphinx 1.2+ by not requiring it in conf.py.in; we were
not using it before anyway (the `refcount_file' variable was not set) and it
was likely added by default when the configuration file was generated.

Change-Id: I6c775fb4c1c2daae25aade68079efc2e77225a0f
Reviewed-by: John Ehresman <jpe@wingware.com>