Return enums in declaration order (order added)
Modify _ScopeModelItem to return enums (from the enums() method) in the
order that they were added (which presumably is the order in which they
were declared). We must do this because we must process enumerations in
the same order in order to resolve values, as later declared enums may
refer to values from earlier declared enums (and in fact, this is
exactly the case in the 'testenum' test), and the order we get just from
QHash may not match declaration order.

Change-Id: I15a05df98a2cee7ecccb6c82d3f9017735281245
Reviewed-by: John Cummings <jcummings2@users.sf.net>