Fix segfault when disconnecting signal on deleted object
When disconnecting a signal on object that has been destroyed the
disconnect call (Sbk_QObjectFunc_disconnect) fails and returns 0 with an
error set. The calling function (signalInstanceDisconnect) was
segfaulting because it decrements the reference count of the returned
value. As the Python error is already set it's sufficient for
signalInstanceDisconnect to return 0 in this case.

Change-Id: If857961c0d64360045c1f2ba5855881e2a9aae43
Task-number: PYSIDE-189