finishing this explorative branch with success.
The problem with missing wrappers was in global.h.in . This file was changed
to use "@QT_QTXXX_INCLUDE_DIR@/xxx.h"  macros all the time instead of
using <QtXXX/xxx.h> . The new Qt5XXX_INCLUDE_DIRS variables can not simply be used,
because they are lists of folders.
I cannot find the reason for this change, but will revert this,
in the hope that the new include structure of Qt5 is better than Qt4's.