fixed a problem with qRegisterMetaType that came up with Qt5.
The problem is that an unqualified name gets registered ("iterator") by some hard to
understand heuristics of shiboken. This way, the same name is used for objects which have
different size in Qt5, and therefore things crash.

The quick fix was to skip exactly the name "iterator" and output a warning.
This needs to be fixed in a better way, but costs too much time at the moment,
because I don't understand enough about possible side-effects, yet.