Support QtTest and the special PySideQTouchEventSequence
This was a bit difficult to understand. After I realized, that the code is mostly a copy
of Qt code, things became clearer: The modules corrected a bit in Qt 4.6 and also
turned the sequence into heap-based.
I did the same for Qt5 (the correction was no longer needed).
I tried very hard, but still cannot find out why signature
    'generateTouchEvent(QWindow*,QTouchDevice*,bool)'
is not found. May somebody enlighten me.