fix naming problem with QtWebKit
This was a tricky bug concerning CamelCase and cmake-caching.
cmake has the habit of remembering the case of variables as they
have been used in a script.
The bad effect is that when a wrong case is tried once, cmake remembers this
and no longer finds the variable with the correct case!

I was about to report a bug, when I finally found the cache file.