ensure that a type check is done on possibly unknown primitive types.
This ids about a hard crash in QtMultimedia. Hacked quite a while until
the reason was clear:
The combination of a missing type check and the generally bad "reinterpret_cast"
resulted in this error. I changed the cause to "dynamic_cast" and added an error message.

The problem is not fixed, yet. It only does not core dump, anymore.