get rid of QGraphics enum problems
This was a hardto-find problem when moving to QtWidgets.
The rejection of the unnamed field "enum_1" did not make it
into the type system, and this had a huge effect on compilation.
See the xml file fore explainment.

Things like these are a major distraction from shiboken, because problems
stay very deeply buried. In the end, I only happened to find this by comparing
the log files of a Qt4.8 build pf PySide.