start a final clean-up before switching to Qt 5.5
There are still some warnings generated in the not-so-urgent modules, and
QtWebKit even has an unresolved class leftover. These things need to be totally
clean. Then I will tag that version as something like "clean_5.4", and move on
to 5.5.

If people need to create a 5.4 version, they may go back here and create a branch
or build some conditional compilation. I will not spend time on this during this project phase.