get rid of unresolved warnings in QtTest.
This was bugging me for a long time, until I learnt that things reported as "signature not found"
can have many meanings, not only the function name of the correctness of the arguments,
but also a simple missing declaration of a new type.

Shiboken should be more verbose here. But after you know that...