remove all CMAKE_CURRENT_BINARY_DIR variables in favor of the much clearer ${name}_GEN_DIR variable.
If that did not fit, the CMAKE_CURRENT_XXX was changed to explicit module variables.
This change made a lot of sense, improving readability and reducing the chance of typos.
Small drawback: the use of macro(HAS_QT_MODULE var name) is now mandatory, because
it creates the new variable as a side effect.