make test binding build correctly. One warning left.
This is almost ready, just one warning is still there.
It was crucial to add the "--api-version" parameter to the shiboken command.
Funny, this must have been wrong as long as version numbers are used in the XML!

Is this test binding module not used, or does it work with all the warnings?