silence that message about 'QPyTextObject'
I hate to do that.
But after checking against the Qt 4.8.7 version, this is not any worse.
From the check-in:

"""
I never really understood this warning. Probably it is because there
is no way to instantiate the class. Anyway, why must this class emit this warning?
I am not a C++ warrior, so I'd prefer if it would say "it is ok if you see this warning!".
Well, maybe somebody will enlighten me, and I'll change this comment.
I'd actually prefer an implementation that avoids generating this message, but I still failed
doing so :-(
"""