work around a 'jom' bug on windows.
Om windows, when no sphinx tool is installed, a build with 'jom' would crash,
because the "doc" target is not created. This is a 'jom' bug, because things work with nmake.

By always generating a "doc" target, this is circumvented. Instead of crashing, I write
a message. This is quite helpful for test builds.

Also, a wrong "win32" variable was used, which must be upper case.