Patchset to install Python* extensions correctly (at least on Linux) While updating the old packaging on Ubuntu, I found some (dirty) fixes done by the maintainers of the upstream packaging. So I decided to fix all of that here, to make these fixes unneeded.
1. Since Python3 all Extensions need at least something like {extension}.cpython-35m.so appended to the libraries names.
Depending on an addition in Shiboken2, I added this for all Python3 versions less 3.5.

2. Additionally all Python3 versions higher or equal 3.5 need {extension}.cpython-35m-{MULTIARCH_SUFFIX}
Following the debian documentation I added a short procedure to get this string and append it also to every library name.

3. When targeting "Release" and "Debug" Python2.* extensions are called the same, eg. {extension}.{MULTIARCH_SUFFIX}.so.
Here (and on other python packages) the debug libraries have got "_d" appended - also added this nativly in CMake. As a result, when building now for "Release" and "Debug" it isn't needed to have two different DISTDIR= calls, eg. DISTDIR=debian/tmp and DISTDIR=debian/tmp-dbg, because of overwritten debug libraries.