Fix running the generator during build on Visual Studio.
On Windows, when using Visual Studio, there is an additional "Debug" or "Release" directory under the
`<project_name>`_BINARY_DIR path so running the generator would sometimes fail.

In later versions of CMake, we can pass the executable target's name directly to `add_custom_command`
and it will use the correct path.