Intoducing new option "qt" I would even call it qt_version, but qtchooser, which switches between different versions of Qt, is also using this option. So I hope this would be less confusing for other people. So in case you have different Qt5 installations you can choose now one of them by using --qt=5.4 for example.
```
$ qtchooser -l
4
5
default
qt4-i386-linux-gnu
qt4-x86_64-linux-gnu
qt4
qt5-x86_64-linux-gnu
qt5
```

The main reason for this commit is that qtchooser chooses qt4 as default here on Ubuntu/Linux (as we can also see in the travis builds). This commit will add the option plus add the default to qt5.