Fix compilation problem with VS2015
VS2015 does not build pyside, because the compiler does not allow access
to a private constructor.
Older versions of Visual Studio were more sloppy and had no problem.

On other platforms, the so-called "protected hack" does the trick.

Unfortunately, VS2015 cannot use this hack, because even if enforced,
the linker would not find the symbols, because of extensive name mangling.

One unpopular solution would be to build Qt5 with the same settings.

My current solution is to modify the generated header files by adding
public constructors. This is a temporary solution until we improve
shiboken or even replace it.

[ChangeLog][pyside2][PYSIDE-321]
Fix Windows: Pyside2 does not compile with MSVC2015/Update2

Change-Id: I1e278889be688a00b4dd93abd8c87db1a026f006
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>
Reviewed-by: Christian Tismer <tismer@stackless.com>