Fix the shiboken tests on VS2015
The shiboken tests do not build, so we cannot test at all with VS2015.
One issue was a simple C++11 restriction, the other involves the
classes in photon.cpp / photon.h .

I could not find out how to fix that, so I removed the photon files,
in order to get a running test framework.

If somebody knows an easy patch instead of removing photon, please
let me know. Otherwise I would prefer pretty much to move on.

[ChangeLog][shiboken][PYSIDE-324]
Fix compilation of tests with VS2015

Change-Id: I6a8d64b18c5eba110f4c03dba63c4b21d6816918
Reviewed-by: Christian Tismer <tismer@stackless.com>