Fix the shiboken tests on VS2015
The shiboken tests do not build, so we cannot test at all with VS2015.
One issue was a simple C++11 restriction, the other involves the
classes in photon.cpp / photon.h .

The other issue was a problem with the template classes.
This will be fixed in the next commit.

[ChangeLog][shiboken][PYSIDE-324]
Fix compilation of tests with VS2015

Change-Id: I1a04219162b71cf5a20e9134543fa71aa3bc2f92
Reviewed-by: Christian Tismer <tismer@stackless.com>