Fix DynamicQMetaObject property type registration
Previously if a property was defined in python code, libPySide would
register the wrong type id inside the QMetaObject.d.data struct,
specifically it would straight away try to register the type as a
*new* type, without checking if it is a built-in type.
This led to setters not working on the respective property, because
the QVariant which stored the new value, could not be converted to
the type which QMetaObject reported.

Fix consists in checking if the specified property type is a built-in
one, and use it, or register the type as a new one, and make sure
it is tagged as a so-far unresolved type, which will be resolved
later upon access of the property.

Change-Id: Ie5e78a5e760682ce5a98dea138b16354b51c17ae
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Christian Tismer <tismer@stackless.com>