support pyside CI testing with filtering
PySide is currently in an unfinished state. Therefore, running all the
tests will never return a success.

To make this more useful, this script is controlled by a blacklist,
in order to keep record of the known-to-fail tests.

On success: zero error status
On Failure: nonzero (raises ValueError)

The test parser uses a blacklist file that is compatible to standard
Qt blacklist files.

macOS, Ubuntu and windows (thanks to fkleint) are supported.

The blacklist has a feature where new configurations are learned.
It first removes old versions of the same tests, before it inserts the new ones.

Blacklisting is now optional.
We support now py2, py3, qt5, qt5.6 etc.

Some consideration about collapsing the data are added.
The parser has now a sub-command "test" that is mandatory.

Support for moving the build_dir to another location without loosing the tests so far.

Implement BPASS.
Support CTEST_OUTPUT_ON_FAILURE

Change-Id: If66d3cf1733a3b0c8a47fe9b8c6aec8ea430c699
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>