Restructure setup.py
Move all code that was unconditionally executed at startup
into a subroutine prepareBuild() and execute that from the build command
class.

Overwrite the Distutils'_install to be extended.

This makes the standard --help options of DistUtils work as expected
and causes the script to require a command argument instead of
starting unconditionally.

Change-Id: I0bf1bfc4ba0e15662122733b04e468f1838e4ae8
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>