GlobalReceiverV2: Fix double delete of receivers on exit
When SignalManager is being destructed and calls qDeleteAll on the
map of global receivers, a deletion of one receiver might indirectly
cause another receiver to be deleted. This will cause a double
delete when qDeleteAll tries to delete the already deleted second
receiver.

Fix this by deleting the first map element in a loop while the map
is not empty. This makes sure that if an indirect delete does happen,
it won't invalidate any iterator, because we call begin() again to
get a new iterator.

Task-number: PYSIDE-88
Change-Id: I9e812271c1a86607094006bc01eb55c930d7fa95
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>