fix testing with filtering on real CI
There are a couple of effects when running testrunner.py in real life.

Fix: It reports success when there are no tests found.

Fix and extend the recognition of relative paths.

Add a "getcwd" sub-command to get external access to the build folder.

The make command uses no longer "-C" for easier environ access.

Change-Id: Iee2f81ed762b9773021055d26f9b144c31acbe35
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>