Make paths absolute in setup.py
The paths for "qmake" and "cmake" will not work when they are relative paths.

This problem is pretty subtle, when setup.py breaks because it cannot create the
simple list at line 712, variable "cmake_cmd". This innocent looking list is causing problems, because
it uses qtinfo.py, and this is dependent on properties which call back into
the subprocess module!

The properties in qtinfo.py are now real properties, because their values are
early computed in __init__. The problem is solved by this patch.

Change-Id: I877b6644fa2909ca9ac1f23d4ce5accfc869716b
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>