Fix property notify ids to be correct
When Python parses a class to find signals and slots, the resulting
list contains the methods in the order they are defined in the source
code. After that, properties are parsed, and the notify ids are set
based on the relative order of the parsed methods.
But because the methods are sorted before their details are put into
the dynamic QMetaObject, the notify id of a property might point to
a wrong method.

This fix makes sure that the relative notify id of every property is
recomputed before being written to the QMetaObject. This allows
QML bindings to work correctly.

Change-Id: I5e50970fee9a42ae8ea551b7891d396b8a6eda97
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>