Fix shiboken tests from being rebuilt if file contents didn't change
Previously when using setup.py with --reuse-build, very often the
shiboken tests got rebuilt, even though nothing changed in the
generated wrapper files.
The problem was due to shiboken not updating the last modified
timestamp of generated files, in case if the new code was identical
to the previous one. This led to cmake continuously re-running the
generation of files, and subsequent compilation.

This fix makes sure to change the last modified timestamp of each
generated file, even if the contents didn't change, thus stopping
the wasteful rebuilds.

Change-Id: I5ca1791c12133b34a5d4c510e5c91ab288b8ddc6
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>